Move data quality DAG creation into the main DAG file
